feat: Add yaru dependency and simplify pubspec.yaml
This commit adds the `yaru` package as a dependency to the project. The Yaru package provides Ubuntu's Yaru theme widgets for Flutter applications.

The `pubspec.yaml` file has been simplified by removing many of the boilerplate comments.

The `pubspec.lock` file has been updated to reflect the addition of `yaru` and its transitive dependencies, as well as updates to existing dependencies due to the `flutter pub get` command.